requirements.txt
annotated-types==0.7.0
casaconfig==1.0.2
casatasks==6.6.5.31
casatools==6.6.5.31
certifi==2024.12.14
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
contourpy==1.3.1
cycler==0.12.1
docker-pycreds==0.4.0
fastprogress==1.0.3
fonttools==4.55.3
gitdb==4.0.11
GitPython==3.1.43
idna==3.10
kiwisolver==1.4.7
mantaray-client==1.2.6
matplotlib==3.10.0
numpy==1.26.4
packaging==24.2
pillow==11.0.0
platformdirs==4.3.6
protobuf==5.29.2
psutil==6.1.0
pydantic==2.10.4
pydantic_core==2.27.2
pyerfa==2.0.1.5
pyparsing==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
PyYAML==6.0.2
requests==2.32.3
rootutils==1.0.7
scipy==1.14.1
sentry-sdk==2.19.2
setproctitle==1.3.4
six==1.17.0
smmap==5.0.1
typing_extensions==4.12.2
urllib3==2.2.3
wandb==0.19.1
websocket-client==1.8.0
